fix(lifecycle): move lifecycle function into separate files to support TS
