+++ src/util/lifecycle-rendered.js
@@ -1,0 +1,1 @@
+export default function rendered() {}
